Make autoparagraph on by default
Set the default to yes. Use the pragma for the tiddler so that it is set for parsing the tiddler and then allow it to be set by the config for child transclusions. This means that the default is to continue like normal so there are no backwards-incompatible changes.

The view template body, and the model both reset this parameter to yes to allow theme developers to specify everything outside of the normal tiddler view template to be not autoparagraphed, and developers may specify anywhere they like that they want autoparagraph turned on or off simply by using the set widget.