Unify behaviour of standalone elements in block contexts
Standalone elements in block contexts are now always parsed in block mode. Removes for example this gotcha:

\define foo()
<button/>
\end

<<foo>>

Becomes before this commit: <p><button></button></p> (because the \end consumes the newline before)
Becomes after this commit: <button></button> (because 'end of input' now also counts as linebreak)

To have the button wrapped in a p, either do ...

\define foo()
<button></button>
\end

... or ...

\define foo()
<button/>
\end

<$macrocall $name='foo'></$macrocall>

(the closing tag of the macrocall generates the p)

I already mentioned this topic in the code example under 5 in my PR message.

Due to this change it is also now irrelevant if a tiddler ends with a line break or not. Which before also could sometimes lead to unexpected p wrapping. For example:

<img>

<img>

<img>(end of tiddler, no line break)

Becomes before this commit: <img><img><p><img></p> (because there is no line break after the last img)
Becomes after this commit: <img><img><img> (because 'end of input' now also counts as line break)

Notice, this means that there is currently no way to have HTML5 void elements standing alone (like the first to <img>es above) be automatically wrapped in a p. One needs to explicitely wrap them with a p. Before one could transclude them from tiddlers having no line break at the end, or define them as macros like in the first example. But in my opinion the intuitive behaviour, introduced by this commit outweights that small disadvantage.

Also, in general, we should not rely on if a tiddler ends with a newline or if there is an empty line before an \end .