Example reformatting of the wikitext-example macro
Applied only on editions/tw5.com

notice the 'dn' in front of the attribute values:

* d: dedent
* n: strip newlines from start, whitespaces from end
* s: strip whitespaces from start and end

These get executed in the order they are defined, so nd"..." can result in a different
result than dn"...".

The feature is currently not implemented for macro definition parameters.

Should we make '''...''' be an alias for dn"""...""" ?