Remove unnecessary newlines before \end
'\end's consume the linebreak before them.

Until now that would mean that standalone elements or opening tags which miss their closing tag would be parsed in inline mode when there is no empty line between them and the '\end'. Same true when they are at the end of a Tiddler, with no newline following.

This is confusing for users.

Therefore i made 'end of input' also count as newline (in the commit 'unify behaviour of standalone elements in block contexts').

So now these newlines before '\end's are not needed any more.