Delete PETTIL example
The domain of the linked site, http://chitselb.com/files/tiddlypettil.html, has expired and now links to https://www.hugedomains.com/domain_profile.cfm?d=chitselb&e=com. Note that the linked source repository is still available https://github.com/chitselb/pettil

I'm not sure whether it's better to delete the whole example because the wiki is not available without rehosting or to leave it there for reference but I chose the former.

afaict, deleting examples from this page will remove them from the listing because it's aggregated via tags but I'm not sure and I have not tested this change because I don't know how and I'm short on time.