Fix blank favicon if root-tiddler=save-lazy-*
If the `root-tiddler` option is set to `save-lazy-images` or
`save-lazy-all` on node.js the favicon is empty because the
tiddler `$:/favicon.ico` that is used to source the icon will be lazy
loaded and is empty at first. The favicon is show as soon as the
tiddler is opened.

This patch excludes the `$:/favicon.ico` tiddler from lazy loading and
restores the favicon functionality.