Move compression into sendResponse to compress all responses
All responses are funneled through `sendResponse` to handle the caching.
Centralising `deflate` or `gzip` compression there as well has two big
benefits:
- Compression is only done if the response will actually be sent
- Compression can be used for all handlers (for eg. lazily loaded
  images)

The compression is only done if the data to send is more than 2048 bytes
because the bandwidth saving is neglectible below that threshold and the
preformance cost of compression and decompression might get significant
at that point.

The following table lists the download size with the root-tiddler set to
`$:/core/save/all-external-js`:

|                    |   gzip=no   |   zip=yes   |
|--------------------|-------------|-------------|
| Without this patch | 3,04 MiByte | 2,43 MiByte |
| After this patch   | 3,04 MiByte |  700 kiByte |