Implement no-wrap syntax
We are at the end of an opening tag. We check for one line break (previously two). If there is one, this element will be parsed with the block html rule, and therefore will not be wrapped in a paragraph. Otherwise it will (later) be parsed with the inline html rule, and paragraph wrapping will happen.