Rename variable
We are still at the end of the same opening tag, like in the previous commit. That code location has already been called during parsing.

This time we check if there are TWO line breaks after the opening tag.

If so, the CHILD will be parsed with a block rule, which means, it may be wrapped in a p.

Otherwise it will be parsed with an inline rule, which means it will not be wrapped in a p.