Implement escaped line break syntax
A backslash ('\') directly after an opening tag escapes a directly following linebreak. The backslash and the linebreak are thrown away.

So we can write

<span>\
...</span>

and this gets parsed as if we had written

<span>...</span>

aka, this span will get wrapped in a paragraph. Without the backslash, the newline would tell the parser to not wrap in a paragraph, due to the changes introduced by the first commit.