Implement trimming of newlines at the start of pre and code elements
As a newline after the opening tag is now sometimes needed to prevent wrapping in a paragraph, we need to trim away newlines at the start of pre and code elements, to prevent that they have empty lines at their start when viewed in the browser.

We leave other whitespace unchanged, like indents and whitespace between text.