Close paragraph before calling a macro
... by adding an empty line before the macro call. Because the macro call resolves to a div and we dont want to have that wrapped in a paragraph.

Play around with this code example ...

    \define foo()
    <div>|hi</div>
    \end

    hi|<<foo>>|hi

... in a tiddler created at https://nilslindemann.github.io/TiddlyWiki5-unified-html-syntax--PR-Work-Folder/singlefile.html, and add one or two newlines at the locations where i added a '|', to figure out how this behaves in regard to paragraph wrapping.

You will see that we need two newlines between text and a following macro call, in order to have the text wrapped in a paragraph.

On the other hand we only need one newline between a macrocall and a following text, for the text to be wrapped in a paragraph.

This is due to the way $:/core/modules/parsers/wikiparser/rules/macrocallblock.js -> findNextMatch() works (which my PR does not touch). There we check for a single line break after the macro call, to detect if this is a block macro call or an inline macro call. In the first case, text after the line break gets wrapped in a paragraph, in the second it doesn't.

This can be unintuitive. For example:

    \define foo() b

    a
    b
    c

    a
    <<foo>>
    c

    a
    b
    c

In the browser this gets displayed as three paragraphs, each containing "a b c".

If we remove the "c" directly after the macro call, we still get three paragraphs, "a b c", "a b", and "a b c".

If we re-add the "c" and remove the "a" directly before the macro call, we suddely get four paragraphs, "a b c", "b", "c", and "a b c".

If instead, in above javascript file, we check for two newlines instead of just one, we would get "a b c", "b c", and "a b c" in the last case.

In my opinion, this would be more in sync with other WikiText block elements, which in other locations require two newlines before them, but currently not after macro calls.

We would also not have to add invisible whitespaces after macro calls, in order to have them parsed inline. This can easily be overlooked - the following commit in this PR, "Remove wrapping paragraph caused by whitespace after macro call" fixes one case. And these spaces can also accidently be stripped away by editors, which trim line ends on saving.

This is another syntax change we could include, on the day we merge the syntax introduced by this PR into the master. If we decide to merge it. And depending on the way we do that exactly, to preserve backwards compatibility. See Jeremys mime type or vocabulary parameter suggestion – https://github.com/Jermolene/TiddlyWiki5/pull/6135#issuecomment-950128144. Or my "parsing palettes" suggestion – https://github.com/Jermolene/TiddlyWiki5/pull/6135#issuecomment-955046909.