Wrap two link widgets in paragraphs
... by adding a newline before them.

Because the wrapping set widget does not get wrapped in a paragraph any more.

Here the same effect can be achieved by e.g. adding a line break escaper after the opening tag of the set widget. But if we then add more link widgets inside, they will all together be wrapped in a paragraph, not each of them.