Add type coercion for HTML5 date input
This patch adds infrastructure to do type coercion on input fields. It
introduces a new module-type `inputtype` that can convert values between
the internal TW5 representation and the representation used by the value
attribute of the HTML-tag.

Modules of type `inputtype` export a `type` property that specifies on
which type of `input` elements they used. In addition to that
they contain two exported functions `toValue` and `fromValue`. These
functions are called each time a value needs to be converted from TW5`s
representation into the HTML-value-attribute and vice versa.

`toValue` is used to convert from TW5-value to HTML-value-attribute. The
TW5-value is passed as its first parameter. The function must return
the value used for the HTML input tag's value-attribute.

`fromValue` is used to convert from HTML-value-attribute to TW5-value.
The HTML-value is passed as its first parameter. The function must
return the value that should be used by TW5.

Currently inputtype-modules are only defined for the types `date` and
`datetime-local`. These types are supported by most mainstream browsers.
Adding the `week` and `month` types would be easy if this is desired.